feat: split verifiable field components into own feature directory
