+++ frontend/src/features/verifiable-fields/components/VerifiableFieldContainer/index.ts
@@ -1,0 +1,2 @@
+export type { VerifiableFieldContainerProps } from './VerifiableFieldContainer'
+export { VerifiableFieldContainer } from './VerifiableFieldContainer'
