+++ frontend/src/features/verifiable-fields/components/VerificationBox/MobileOtpSvgr.tsx
