+++ frontend/src/features/verifiable-fields/components/VerificationBox/index.ts
