+++ frontend/src/features/verifiable-fields/types.ts
@@ -1,0 +1,13 @@
+import type { EmailFieldBase, MobileFieldBase } from '~shared/types/field'
+
+export type VerifiableFieldBase = MobileFieldBase | EmailFieldBase
+export type VerifiableFieldType = VerifiableFieldBase['fieldType']
+
+export type VerifiableFieldSchema<T extends VerifiableFieldBase> = T & {
+  isVerifiable: true
+}
+
+export type VerifiableFieldInput = {
+  value?: string
+  signature?: string
+}
