--- frontend/src/templates/VerifiableField/index.ts
@@ -1,12 +1,0 @@
-export * from './types'
-export type { VerifiableFieldContainerProps } from './VerifiableFieldContainer'
-export { VerifiableFieldContainer } from './VerifiableFieldContainer'
-export {
-  useVerifiableField,
-  VerifiableFieldContext,
-} from './VerifiableFieldContext'
-export type {
-  VerifiableFieldProps,
-  VerifiableFieldProviderProps,
-} from './VerifiableFieldProvider'
-export { VerifiableFieldProvider } from './VerifiableFieldProvider'
